[Fix] #1 coderabbitai 코드 리뷰 반영

+++ src/main/java/com/nova/yeobaek/domain/item/domain/Category.java
@@ -43,7 +43,7 @@ public class Category extends BaseEntity {
 	private String name;
 
 	@Column(nullable = false)
-	private Long level;
+	private Integer level;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "parent_id")
